Fixed warnings
Also fixed the editor camera being updated even when the scene is playing

+++ Debut/src/Debut/Renderer/SubTexture2D.cpp
@@ -5,8 +5,8 @@ namespace Debut
 {
 	Ref<SubTexture2D> SubTexture2D::CreateFromCoords(const Ref<Texture2D>& texture, const glm::vec2& coords, const glm::vec2& spriteSizeInTiles, const glm::vec2& tileSize)
 	{
-		float sheetWidth = texture->GetWidth();
-		float sheetHeight = texture->GetHeight();
+		uint32_t sheetWidth = texture->GetWidth();
+		uint32_t sheetHeight = texture->GetHeight();
 		
 		glm::vec2 min = glm::vec2((float)(coords.x * tileSize.x) / sheetWidth, (float)(coords.y * tileSize.y) / sheetHeight);
 		glm::vec2 max = glm::vec2((float)((coords.x + spriteSizeInTiles.x) * tileSize.x) / sheetWidth, (float)((coords.y + spriteSizeInTiles.y) * tileSize.y) / sheetHeight);
